+++ src/Codec/CBOR/Cuddle/CBOR/Validator.hs
@@ -21,7 +21,6 @@ import Codec.CBOR.Cuddle.CBOR.Validator.Trace (
   evidence,
   isValid,
   mapTrace,
-  showSimple,
  )
 import Codec.CBOR.Cuddle.CDDL hiding (CDDL, Group, Rule)
 import Codec.CBOR.Cuddle.CDDL.CTree
@@ -36,6 +35,7 @@ import Codec.CBOR.Cuddle.CDDL.Custom.Validator (
   XXCTree (..),
   runValidator,
  )
+import Codec.CBOR.Cuddle.CDDL.Resolve (showSimple)
 import Codec.CBOR.Cuddle.IndexMappable (IndexMappable (..))
 import Codec.CBOR.Read
 import Codec.CBOR.Term